refactor(src/child-process.js): rename for clarity, gonna make this a lil' grab bag
